pyproject.toml
[project]
name = "starfield-tool"
version = "0.1.0"
description = "Modular GUI toolkit for Bethesda's Starfield"
requires-python = ">=3.12"
dependencies = [
    "customtkinter>=5.2",
    "watchdog>=4.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/starfield_tool", "src/bethesda_creations", "src/load_order_sorter"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.15.8",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "."]
